+++ src/js/integrations/default.ts
@@ -1,25 +1,34 @@
-import { HttpClient } from '@sentry/integrations';
-import { Integrations as BrowserReactIntegrations } from '@sentry/react';
 import type { Integration } from '@sentry/types';
 
 import type { ReactNativeClientOptions } from '../options';
-import { HermesProfiling } from '../profiling/integration';
 import { ReactNativeTracing } from '../tracing';
 import { isExpoGo, notWeb } from '../utils/environment';
-import { DebugSymbolicator } from './debugsymbolicator';
-import { DeviceContext } from './devicecontext';
-import { EventOrigin } from './eventorigin';
-import { ExpoContext } from './expocontext';
-import { ModulesLoader } from './modulesloader';
-import { NativeLinkedErrors } from './nativelinkederrors';
-import { ReactNativeErrorHandlers } from './reactnativeerrorhandlers';
-import { ReactNativeInfo } from './reactnativeinfo';
-import { Release } from './release';
+import {
+  breadcrumbsIntegration,
+  browserApiErrorsIntegration,
+  browserGlobalHandlersIntegration,
+  browserLinkedErrorsIntegration,
+  debugSymbolicatorIntegration,
+  dedupeIntegration,
+  deviceContextIntegration,
+  eventOriginIntegration,
+  expoContextIntegration,
+  functionToStringIntegration,
+  hermesProfilingIntegration,
+  httpClientIntegration,
+  httpContextIntegration,
+  inboundFiltersIntegration,
+  modulesLoaderIntegration,
+  nativeLinkedErrorsIntegration,
+  nativeReleaseIntegration,
+  reactNativeErrorHandlersIntegration,
+  reactNativeInfoIntegration,
+  screenshotIntegration,
+  sdkInfoIntegration,
+  spotlightIntegration,
+  viewHierarchyIntegration,
+} from './exports';
 import { createReactNativeRewriteFrames } from './rewriteframes';
-import { Screenshot } from './screenshot';
-import { SdkInfo } from './sdkinfo';
-import { Spotlight } from './spotlight';
-import { ViewHierarchy } from './viewhierarchy';
 
 /**
  * Returns the default ReactNative integrations based on the current environment.
@@ -33,47 +42,47 @@ export function getDefaultIntegrations(options: ReactNativeClientOptions): Integ
 
   if (notWeb()) {
     integrations.push(
-      new ReactNativeErrorHandlers({
+      reactNativeErrorHandlersIntegration({
         patchGlobalPromise: options.patchGlobalPromise,
       }),
     );
-    integrations.push(new NativeLinkedErrors());
+    integrations.push(nativeLinkedErrorsIntegration());
   } else {
-    integrations.push(new BrowserReactIntegrations.TryCatch());
-    integrations.push(new BrowserReactIntegrations.GlobalHandlers());
-    integrations.push(new BrowserReactIntegrations.LinkedErrors());
+    integrations.push(browserApiErrorsIntegration());
+    integrations.push(browserGlobalHandlersIntegration());
+    integrations.push(browserLinkedErrorsIntegration());
   }
 
   // @sentry/react default integrations
-  integrations.push(new BrowserReactIntegrations.InboundFilters());
-  integrations.push(new BrowserReactIntegrations.FunctionToString());
-  integrations.push(new BrowserReactIntegrations.Breadcrumbs());
-  integrations.push(new BrowserReactIntegrations.Dedupe());
-  integrations.push(new BrowserReactIntegrations.HttpContext());
+  integrations.push(inboundFiltersIntegration());
+  integrations.push(functionToStringIntegration());
+  integrations.push(breadcrumbsIntegration());
+  integrations.push(dedupeIntegration());
+  integrations.push(httpContextIntegration());
   // end @sentry/react-native default integrations
 
-  integrations.push(new Release());
-  integrations.push(new EventOrigin());
-  integrations.push(new SdkInfo());
-  integrations.push(new ReactNativeInfo());
+  integrations.push(nativeReleaseIntegration());
+  integrations.push(eventOriginIntegration());
+  integrations.push(sdkInfoIntegration());
+  integrations.push(reactNativeInfoIntegration());
 
   if (__DEV__ && notWeb()) {
-    integrations.push(new DebugSymbolicator());
+    integrations.push(debugSymbolicatorIntegration());
   }
 
   integrations.push(createReactNativeRewriteFrames());
 
   if (options.enableNative) {
-    integrations.push(new DeviceContext());
-    integrations.push(new ModulesLoader());
+    integrations.push(deviceContextIntegration());
+    integrations.push(modulesLoaderIntegration());
     if (options.attachScreenshot) {
-      integrations.push(new Screenshot());
+      integrations.push(screenshotIntegration());
     }
     if (options.attachViewHierarchy) {
-      integrations.push(new ViewHierarchy());
+      integrations.push(viewHierarchyIntegration());
     }
     if (options._experiments && typeof options._experiments.profilesSampleRate === 'number') {
-      integrations.push(new HermesProfiling());
+      integrations.push(hermesProfilingIntegration());
     }
   }
 
@@ -88,16 +97,16 @@ export function getDefaultIntegrations(options: ReactNativeClientOptions): Integ
     integrations.push(new ReactNativeTracing());
   }
   if (options.enableCaptureFailedRequests) {
-    integrations.push(new HttpClient());
+    integrations.push(httpClientIntegration());
   }
 
   if (isExpoGo()) {
-    integrations.push(new ExpoContext());
+    integrations.push(expoContextIntegration());
   }
 
   if (options.enableSpotlight) {
     integrations.push(
-      Spotlight({
+      spotlightIntegration({
         sidecarUrl: options.spotlightSidecarUrl,
       }),
     );